requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-restframework==0.0.1
django-sslserver==0.22
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
et_xmlfile==2.0.0
google-auth==2.40.3
gunicorn==23.0.0
h11==0.16.0
html2text==2025.4.15
html5lib==1.1
idna==3.10
lxml==5.4.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pypdf==5.6.0
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
reportlab==4.4.1
requests==2.32.3
rsa==4.9.1
setuptools==80.7.1
six==1.17.0
sqlparse==0.5.3
stripe==12.4.0
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
uvicorn==0.35.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
